requirements.txt
jinja2
PyYAML
pymongo
python-dateutil
M2Crypto
cherrypy
pytz
hashlib
markupsafe
dpkt
pypcap
netaddr
pyzmq
yara
